Added working code to compute HalfMatrix SSM and added gitignore for cache files
