+++ .gitignore
@@ -1,0 +1,2 @@
+.cache-noteaholic
+__pycache__/
