+++ Test.py
@@ -1,4 +1,6 @@
-
 # I want to calculate the SSM using the pitches. I want to start from the high level, and calling unimplemented functions is ok. 
 
 # 1. Let's sign into Spotify and get the audio analysis for a song.
+
+# Currently: 
+# Create functions to Gaussian blur the SSM in the compute_harmonic_structure function.
